refactor(canonical-json): single shared JCS canonicalizer for audit + plan hashes
`state/audit.ts::canonicalize` and `api/plan/engine.ts::stableStringify`
were two copies of JCS-style canonical JSON (recursive key-sort, no
whitespace) meant to be identical but already diverged — plan's lacked the
`!(v instanceof Buffer)` guard. Both feed sha256 hashes that must be
deterministic (audit hash chain; plan input_hash/plan_hash), so two
"canonical JSON" functions that disagree is a latent correctness risk.

Extract one layer-neutral helper at lib/canonical-json.ts (the home the
biome import-boundary rule already points to) with the SUPERSET behavior
(keeps the Buffer guard). audit.ts imports + re-exports it; plan/engine.ts
imports it and drops the local copy. The guard is inert for plan's
Buffer-free inputs, so hashes stay byte-identical — pure refactor, no
observable behavior change. Verified: audit + plan-engine suites (23) and
the full suite (603) green, tsc/lint/format clean.
